Add user to event after accepted the invitation

+++ src/main/java/hu/tamas/university/repository/ParticipateInEventRepository.java
@@ -3,7 +3,7 @@ package hu.tamas.university.repository;
 import hu.tamas.university.entity.ParticipateInEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParticipateInEventRepository extends JpaRepository<ParticipateInEvent, Long> {
+public interface ParticipateInEventRepository extends JpaRepository<ParticipateInEvent, Integer> {
 
 	ParticipateInEvent findParticipateInEventById(int id);
 }
